+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Changed
+
+- Ensured that any slurm job that would round down to zero usage would be billed
+  at least 1 node second. This ensures that super small jobs are still
+  billed, and won't round down to zero usage.
+
 ## [0.22.3] - 2026-03-04
 
 ### Changed